14Sprint ver1

+++ src/main/java/ru/practicum/shareit/user/controller/UserController.java
@@ -31,17 +31,17 @@ public class UserController {
     }
 
     @PatchMapping("/{id}")
-    public UserDto updateUser(@PathVariable int id, @RequestBody UserDto userDto) {
+    public UserDto updateUser(@PathVariable Long id, @RequestBody UserDto userDto) {
         return userService.updateUser(id, userDto);
     }
 
     @GetMapping("/{id}")
-    public UserDto getUser(@PathVariable int id) {
+    public UserDto getUser(@PathVariable Long id) {
         return userService.getUserById(id);
     }
 
     @DeleteMapping("/{userId}")
-    public void removeUser(@PathVariable int userId) {
+    public void removeUser(@PathVariable Long userId) {
         userService.removeUserById(userId);
     }
